Updating and cleaning repo
